Generate wall-components more reliably
Before, we had to decide whether to let Marching Squares
start at the upper left or lower right corner of connected walls to produce
the outline. This would differ based on the map at hand, which would
require to trace the inner or outer outline respectively.
MS now just does both and produces twice the walls, of which some are
simply duplicates of each other (so they don't matter, since they're
rendered on top of each other) or distinct traces for inside and outside
are produced.