Kind of fix characters glitching into walls
The underlying issue is related to the melonjs
coordinate system which uses the center point of
a sprite as the origin of an object. Since the DB
uses the upper left corner, this allows characters to
move into the walls a bit. This is partially fixed
by making the hitboxes of all actors slightly larger.
Works mostly fine for ghosts, but still doesn't look
perfect. It's a compromise between looks and having to
perfectly aim your character when trying to enter
a narrow corridor.